pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cadmium-lake"
version = "0.1.0"
description = "Provenance-first cadmium data lake and exploration layer"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Trhova" }
]
dependencies = [
  "beautifulsoup4>=4.12",
  "duckdb>=1.1",
  "httpx>=0.27",
  "lxml>=5.3",
  "matplotlib>=3.9",
  "openpyxl>=3.1",
  "pandas>=2.2",
  "pint>=0.24",
  "polars>=1.17",
  "pydantic>=2.9",
  "pyarrow>=18.0",
  "pyreadstat>=1.2",
  "pyyaml>=6.0.2",
  "tenacity>=9.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "ipykernel>=6.29",
  "jupyter>=1.1",
  "pytest>=8.3",
]
pdf = [
  "camelot-py>=1.0",
  "pdfplumber>=0.11",
  "tabula-py>=2.10",
]

[project.scripts]
cadmium-lake = "cadmium_lake.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
